# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "cybersource",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "variables": {}
    }
  ],
  "paths": {
...
    "/api/v1/aggregators/login": {
      "post": {
        "tags": [
          "aggregators"
        ],
        "summary": "login aggregator",
        "operationId": "loginaggregator",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhZ2dyZWdhdG9ySWQiOjEsInJvbGUiOiJhZ2dyZWdhdG9yIiwiaWF0IjoxNTkxNjk1NjY1LCJleHAiOjE1OTE3ODIwNjV9.bpHDsdLfCvgJ7rNU5JjZwqMviO1EEkniVACav9i6ZaY"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginaggregatorrequest"
              },
              "example": {
                "email": "socialpay@gmail.com",
                "password": "DR5SyK101ju24Xqe1w91wg2NYIMErrqW9bi7MuRhIH4=P3YFf/yhT5m+w3eZMjLxEw==6ZjF9gau62Ywlf4HeS43oQ==",
                "digestId": "44"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": []
      }
    },
...
    "/api/v1/merchants/login": {
      "post": {
        "tags": [
          "merchants"
        ],
        "summary": "login merchant",
        "operationId": "loginmerchant",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginmerchantrequest"
              },
              "example": {
                "email": "merchant_1@email.com",
                "password": "DR5SyK101ju24Xqe1w91wg2NYIMErrqW9bi7MuRhIH4=P3YFf/yhT5m+w3eZMjLxEw==6ZjF9gau62Ywlf4HeS43oQ==",
                "digestId": "44"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/api/v1/admin/login": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "login admin",
        "operationId": "loginadmin",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginadminrequest"
              },
              "example": {
                "username": "valid_username",
                "password": "DR5SyK101ju24Xqe1w91wg2NYIMErrqW9bi7MuRhIH4=P3YFf/yhT5m+w3eZMjLxEw==6ZjF9gau62Ywlf4HeS43oQ==",
                "digestId": "44"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "loginaggregatorrequest": {
        "title": "loginaggregatorrequest",
        "required": [
          "email",
          "password",
          "digestId"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "digestId": {
            "type": "string"
          }
        },
        "example": {
          "email": "socialpay@gmail.com",
          "password": "DR5SyK101ju24Xqe1w91wg2NYIMErrqW9bi7MuRhIH4=P3YFf/yhT5m+w3eZMjLxEw==6ZjF9gau62Ywlf4HeS43oQ==",
          "digestId": "44"
        }
      },
...
      "loginmerchantrequest": {
        "title": "loginmerchantrequest",
        "required": [
          "email",
          "password",
          "digestId"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "digestId": {
            "type": "string"
          }
        },
        "example": {
          "email": "merchant_1@email.com",
          "password": "DR5SyK101ju24Xqe1w91wg2NYIMErrqW9bi7MuRhIH4=P3YFf/yhT5m+w3eZMjLxEw==6ZjF9gau62Ywlf4HeS43oQ==",
          "digestId": "44"
        }
      },
...
      "loginadminrequest": {
        "title": "loginadminrequest",
        "required": [
          "username",
          "password",
          "digestId"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "digestId": {
            "type": "string"
          }
        },
        "example": {
          "username": "valid_username",
          "password": "DR5SyK101ju24Xqe1w91wg2NYIMErrqW9bi7MuRhIH4=P3YFf/yhT5m+w3eZMjLxEw==6ZjF9gau62Ywlf4HeS43oQ==",
          "digestId": "44"
        }
      }
    }
  },
...
}